Refine and reorganize
